Second Commit

+++ model.py
@@ -3,7 +3,7 @@ from torch import nn
 import torch
 import matplotlib.pyplot as plt
 from dataset_factory import dataset_factory
-import matplotlib2tikz
+import tikzplotlib
 from pathloss_38901 import pathloss_38901
 import numpy as np
 
@@ -213,8 +213,8 @@ class SkynetModel(nn.Module):
         offset = torch.empty((distance.shape))
         frequency[features[:,7] == 1] = torch.tensor(2.63)
         frequency[features[:,7] != 1] = torch.tensor(0.811)
-        offset[features[:,7] == 1] = torch.tensor(self.offset_2630)
-        offset[features[:,7] != 1] = torch.tensor(self.offset_811)
+        offset[features[:,7] == 1] = torch.tensor(self.offset_2630).float()
+        offset[features[:,7] != 1] = torch.tensor(self.offset_811).float()
         return frequency, offset
 
     def MSE_physicsmodel(self, distance, targets):
